+++ backend/src/main/java/com/upc/brosteria/DTOs/DetallePedidoDTO.java
@@ -3,6 +3,7 @@ package com.upc.brosteria.DTOs;
 import lombok.*;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 @Getter @Setter
 @NoArgsConstructor @AllArgsConstructor
@@ -13,12 +14,12 @@ public class DetallePedidoDTO {
     private Long productoId;
 
     private String productoName;
-    private Double productoPrice;
+    private BigDecimal productoPrice;
 
     @NotNull(message = "La cantidad es obligatoria")
     @Min(value = 1, message = "La cantidad minima de producto debe ser 1")
     private Integer quantity;
 
-    private Double subtotal;
+    private BigDecimal subtotal;
     private String creams;
 }
